requirements.txt
Flask
opencv-python
mediapipe
google-generativeai
numpy
matplotlib
tensorflow
pandas
scikit-learn
scipy
seaborn
keras
torch
torchvision
torchtext
transformers
pytorch-lightning
fastai
jupyter
jupyterlab
jupyterlab-git
jupyterlab-lsp
jupyterlab-plotly
jupyterlab-dash
selenium
webdriver-manager
bs4
langchain
langchain-community 
langchain_huggingface
google-generativeai
faiss-cpu
streamlit
Flask
pymysql
mariadb